% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-test_statistics.R
\name{all_tests}
\alias{all_tests}
\title{Title}
\usage{
all_tests(
  fit,
  svy_design = NULL,
  sim = 1,
  H_inv,
  Delta_mat_list,
  Sigma2,
  pi_tilde
)
}
\arguments{
\item{pi_tilde}{}
}
\description{
Title
}
