% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-test_statistics.R
\name{create_resp_pattern}
\alias{create_resp_pattern}
\title{Create a table of binary response patterns}
\usage{
create_resp_pattern(p = 3)
}
\arguments{
\item{p}{(integer > 0) The number of items.}
}
\value{
A \code{\link[=tibble]{tibble()}} containing ordinal binary values (0/1) for the items as
well as a column indicating one of the \eqn{2^p} response patterns.
}
\description{
Ordering is incremental patterns of power 2 from right to left.
}
\examples{
create_resp_pattern(p = 3)
}
