% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-data_gen_srs.R
\name{gen_data_bin}
\alias{gen_data_bin}
\title{Title}
\usage{
gen_data_bin(
  model.no = 1,
  n = 1000,
  seed = NULL,
  H1 = FALSE,
  return_all = FALSE
)
}
\arguments{
\item{return_all}{}
}
\description{
Title
}
