% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-data_gen_complex.R
\name{gen_data_bin_complex}
\alias{gen_data_bin_complex}
\alias{gen_data_bin_srs}
\alias{gen_data_bin_strat}
\alias{gen_data_bin_clust}
\alias{gen_data_bin_strcl}
\title{Sample from the school population}
\usage{
gen_data_bin_srs(
  population = make_population(1, seed = NULL),
  npsu = 3000,
  seed = NULL
)

gen_data_bin_strat(
  population = make_population(1, seed = NULL),
  npsu = 1000,
  seed = NULL
)

gen_data_bin_clust(
  population = make_population(1, seed = NULL),
  npsu = 140,
  seed = NULL
)

gen_data_bin_strcl(
  population = make_population(1, seed = NULL),
  npsu = 50,
  seed = NULL
)
}
\arguments{
\item{population}{(tibble) Population generated by \code{\link[=make_population]{make_population()}}.}

\item{npsu}{(integer > 0) For the stratified sampling, this is the number of
students within each strata to sample. For the cluster sampling, this is
the number of schools (clusters) to sample. For the stratified cluster
sampling, this is also the number of schools (clusters) per school type
(strata). For the SRS procedure, this is the exact sample size.}

\item{seed}{(numeric) The random seed.}
}
\value{
A \code{\link[=tibble]{tibble()}} containing ordinal binary values (0/1) for the items, as
well as the population stratum and clusters (\code{type}, \code{school}, \code{class}).
}
\description{
There are several functions to simulate a complex (or even
simple random) sampling procedure on the school population generated by
\code{\link[=make_population]{make_population()}}:
\itemize{
\item \code{gen_data_bin_srs()} performs a \strong{simple random sample} on the entire population.
\item \code{gen_data_bin_strat()} performs a \strong{stratified} sampling procedure with the school \code{type} as strata, and the students as the PSU.
\item \code{gen_data_bin_clust()} performs a \strong{two-stage cluster} sampling procedure (ignoring the strata) with \code{school}s being the PSU, and further clustering of \code{class}.
\item \code{gen_data_bin_strcl()} performs a \strong{two-stage stratified cluster} sampling procedure where students are nested within \code{class} within \code{school} (PSU) within \code{type} (strata).
}
}
\examples{
\dontrun{
pop <- make_population(2)
gen_data_bin_srs(pop)
gen_data_bin_strat(pop)
gen_data_bin_clust(pop)
gen_data_bin_strcl(pop)
}


}
\seealso{
\code{\link[=make_population]{make_population()}}
}
