% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32-data_gen_weighted.R
\name{gen_data_bin_wt}
\alias{gen_data_bin_wt}
\title{Generate an informative (weighted) data sample according to \code{model_no}}
\usage{
gen_data_bin_wt(
  model_no = 1,
  seed = NULL,
  H1 = FALSE,
  return_all = FALSE,
  n = 1000
)
}
\arguments{
\item{model_no}{(integer) Choose from 1--5. See pkgdown articles for details.}

\item{seed}{(numeric) The random seed.}

\item{H1}{(logical) Are we generating a data to misspecify the true model?
For power simulations.}

\item{return_all}{(logical) Return the underlying latent variable \eqn{y^*} as
well?}

\item{n}{(numeric > 0) Sample size.}
}
\value{
A \code{\link[=tibble]{tibble()}} containing ordinal binary values (0/1) for the items.
}
\description{
Generate an informative (weighted) data sample according to \code{model_no}
}
\examples{
gen_data_bin_wt(1)
}
