% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.bingof-package.R
\docType{package}
\name{lavaan.bingof-package}
\alias{lavaan.bingof}
\alias{lavaan.bingof-package}
\title{lavaan.bingof: Limited Information Goodness of Fit Tests for Binary Factor Models}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/haziqj/lavaan.bingof}
  \item \url{http://haziqj.ml/lavaan.bingof/}
  \item Report bugs at \url{https://github.com/haziqj/lavaan.bingof/issues}
}

}
\author{
\strong{Maintainer}: Haziq Jamil \email{haziq.jamil@gmail.com} (\href{https://orcid.org/0000-0003-3298-1010}{ORCID})

}
\keyword{internal}
