% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-test_statistics.R
\name{ligof-test-stats}
\alias{ligof-test-stats}
\alias{Wald_test}
\alias{Wald_diag_test}
\alias{Wald_vcovf_test}
\alias{Pearson_RS_test}
\alias{Pearson_test}
\alias{RSS_test}
\alias{Multn_test}
\title{Limited information goodness-of-fit tests}
\usage{
Wald_test(object, approx_Omega2 = FALSE, svy_design = NULL)

Wald_diag_test(object, approx_Omega2 = FALSE, svy_design = NULL, .order = 3)

Wald_vcovf_test(object, svy_design = NULL)

Pearson_RS_test(object, approx_Omega2 = FALSE, svy_design = NULL, .order = 2)

Pearson_test(object, approx_Omega2 = FALSE, svy_design = NULL, .order = "3")

RSS_test(object, approx_Omega2 = FALSE, svy_design = NULL, .order = "3")

Multn_test(object, approx_Omega2 = FALSE, svy_design = NULL, .order = "3")
}
\arguments{
\item{object}{A \code{\link[lavaan:lavaan]{lavaan::lavaan()}} fit object.}

\item{approx_Omega2}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} (logical)  Should an approximate
residual covariance matrix \eqn{\Omega_2} be used? Defaults to \code{FALSE}.}

\item{svy_design}{(optional) A \code{\link[survey:svydesign]{survey::svydesign()}} object.}

\item{.order}{(integer) Either the number of moments to match for the
chi-square test statistic matching procedure (choose from 1--3), or the
Rao-Scott type adjustment order (choose from 1 or 2).}
}
\value{
A data frame containing the test statistics \eqn{X^2}, degrees of
freedom, name of the test, and the \eqn{p}-value.
}
\description{
Limited information goodness-of-fit tests
}
\section{Functions}{
\itemize{
\item \code{Wald_test()}: The Wald test statistic.

\item \code{Wald_diag_test()}: The Wald test statistic using a simple diagonal
\eqn{\Omega_2} matrix.

\item \code{Wald_vcovf_test()}: The Wald test statistic bypassing the
\eqn{\Omega_2} matrix (uses orthogonal complements of \eqn{\Delta_2}).

\item \code{Pearson_RS_test()}: The Pearson test with \eqn{p}-values calculated using a Rao-Scott type adjustment.

\item \code{Pearson_test()}: The Pearson test with \eqn{p}-values calculated using a moment-matching procedure.

\item \code{RSS_test()}: The residual sum of squares (RSS) test. Uses moment-matching for \eqn{p}-value calculations.

\item \code{Multn_test()}: The multinomial test. Uses moment-matching for \eqn{p}-value calculations.

}}
\examples{
fit <- lavaan::sem(txt_mod(1), gen_data_bin(1, n = 500), std.lv = TRUE,
                   estimator = "PML")
Wald_test(fit)
}
\seealso{
\code{\link[=all_tests]{all_tests()}}
}
