% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-sim_functions.R
\name{ligof_sims}
\alias{ligof_sims}
\title{Wrapper function for Type 1 and Power simulations}
\usage{
ligof_sims(
  model_no = 1,
  nsim = 1000,
  samp_size = 1000,
  samp = c("srs", "strat", "clust", "strcl"),
  simtype = c("type1", "power"),
  starting_seed = 4423,
  no.cores = parallel::detectCores() - 2
)
}
\arguments{
\item{model_no}{(integer) Choose from 1--5. See pkgdown articles for details.}

\item{nsim}{(integer) The number of simulations to conduct.}

\item{samp_size}{(integer) The exact sample size for SRS simulations or
stratified sampling; otherwise the \emph{average} sample size for the other
complex sampling methods.}

\item{samp}{(character) Choose the sampling method for the simulated data.
One of \code{srs}, \code{strat}, \code{clust} or \code{strcl}.}

\item{simtype}{(character) Whether this is a \code{type1} simulation or \code{power}
simulation.}

\item{starting_seed}{(integer) The starting random seed.}

\item{no.cores}{(integer) The number of cores to use for parallelisation.}
}
\value{
A list of \code{\link[=tibble]{tibble()}}s with the output from \code{\link[=all_tests]{all_tests()}}.
}
\description{
Wrapper function for Type 1 and Power simulations
}
\examples{
\dontrun{
# To run the SRS Type 1 simulations
analysis_path <- dirname(rstudioapi::getSourceEditorContext()$path)
sim_type <- "type1"
for (the_samp_size in c(100, 250, 500, 1000, 2000, 3000)) {
  for (mod_no in 1:5) {
    sim_name <- paste0("srs", mod_no, "_n", the_samp_size, "_", sim_type)
    cat("[", as.character(Sys.time()), "]", "Now running simulation",
        sim_name, "\n")
    sim <- ligof_sims(mod_no, samp_size = the_samp_size,
                      samp = "srs", simtype = sim_type)
    list2env(setNames(list(sim), sim_name), envir = .GlobalEnv) \%>\% invisible
    save(list = sim_name, file = paste0(analysis_path, "/Rsave/",
                                        sim_name, ".RData"))
  }
}
}
}
