% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_results}
\alias{sim_results}
\alias{res_srs_type1}
\alias{res_srs_power}
\alias{res_complex_type1}
\alias{res_complex_power}
\title{Summarised simulation results}
\format{
A \code{\link[=tibble]{tibble()}} containing the summarised results of the simulation study segregated by test name (Wald, Wald V2, etc.), model (1F 5V, 1F 8V, etc.), and sample size.
\itemize{
\item \code{n_sim} is the number of replications.
\item \code{n_converged} is the number of replications in which \code{\link[lavaan:lavaan]{lavaan::lavaan()}} converged.
\item \code{n_rank_def} is the number of replications where the rank of \eqn{Omega_2} matrix was found to be deficient.
\item \code{rej_rate_x} is the proportions of replications where the null hypothesis was rejected at the \code{x}\% level.
}
}
\usage{
res_srs_type1

res_srs_power

res_complex_type1

res_complex_power
}
\description{
Summarised simulation results
}
\keyword{datasets}
