% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_results_all}
\alias{sim_results_all}
\alias{all_res}
\title{Raw simulation results}
\format{
A list of length 160, each containing a large \code{\link[=tibble]{tibble()}} of raw results.
}
\usage{
all_res
}
\description{
Raw simulation results
}
\details{
For our simulation study, we varied the following settings:
\itemize{
\item Data generation mechanism: \code{srs}, \code{strat}, \code{clust}, \code{strcl}
\item Sample size: 500, 1000, 2000, 3000
\item Factor model: 5 models with various number of items and latent factors. See article details.
\item Simulation type: Type I error analysis or Power analysis
}

Thus, in total there are \eqn{4 \times 4 \times 5 \times 2 = 160} unique simulation scenarios. The R object \code{all_res} is a list of length 160 with each entry containing a \code{\link[=tibble]{tibble()}} of raw results (test statistic value, degrees of freedom, name of test, p-value, convergence, matrix ranks, and random seed) for each scenario. Note that there are 7 tests in total that was investigated, and each scenario was replicated a total of 1000 times.
}
\keyword{datasets}
