% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-transformation_matrices.R
\name{transformation-matrices}
\alias{transformation-matrices}
\alias{create_G_mat}
\alias{create_T2_mat}
\alias{create_Beta_mat}
\title{Create transformation matrices}
\usage{
create_G_mat(p = 3)

create_T2_mat(p = 3)

create_Beta_mat(p = 3)
}
\arguments{
\item{p}{(integer > 0) The number of items.}
}
\value{
A matrix. Additionally, we may inspect the attributes regarding the
ordering of the pairwise components of the \eqn{G} matrix.
}
\description{
The derived limited information test statistics involves some
design matrices which act as transformations from the larger \eqn{2^p}
response pattern space to the lower order univariate and bivariate
marginals.
\itemize{
\item \code{create_G_mat()} returns the \eqn{\tilde R \times R} indicator matrix to obtain all pairwise components.
\item \code{create_T2_mat()} returns the \eqn{p(p+1)/2 \times 2^p} indicator matrix \eqn{T_2} to pick out the unviariate and bivariate moments from the response patterns.
\item \code{create_Beta_mat()} returns the \eqn{4p \times p(p+1)/2} design matrix \eqn{\Beta} described in the manuscript (used to express parameters in terms of residuals).
}

Note that ordering is similar to the ordering in \code{\link[=create_resp_pattern]{create_resp_pattern()}}.
These design matrices currently only apply to binary data. See technical
documents for more details.
}
\examples{
create_G_mat(p = 3)
create_T2_mat(p = 3)
create_Beta_mat(p = 3)
}
\author{
Myrsini Katsikatsou (\code{create_Beta_mat()})
}
