% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utilities.R
\name{txt_mod}
\alias{txt_mod}
\title{Function to return the textual model for lavaan fit}
\usage{
txt_mod(model_no = 1L)
}
\arguments{
\item{model_no}{(integer) Choose from 1--5. See pkgdown articles for details.}
}
\value{
Character vector of factor model to input to \code{\link[lavaan:lavaan]{lavaan::lavaan()}}.
}
\description{
Function to return the textual model for lavaan fit
}
\examples{
txt_mod(1)
txt_mod(2)
txt_mod(3)
txt_mod(4)
txt_mod(5)
}
