% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-test_statistics.R
\name{all_tests}
\alias{all_tests}
\title{Return all test statistics values}
\usage{
all_tests(object, sim = NULL, Sigma2 = NULL, Hinv = NULL)
}
\arguments{
\item{object}{A \code{\link[lavaan:lavaan]{lavaan::lavaan()}} fit object.}

\item{sim}{(integer) Optional and used for large-scale simulations.}

\item{Sigma2}{(for internal testing only)}

\item{Hinv}{(for internal testing only)}
}
\value{
Additionally, if \code{sim} argument is provided, two columns are
appended: Whether the \code{\link[lavaan:lavaan]{lavaan::lavaan()}}  fit has \code{converged} and the
matrix rank of \eqn{\Omega_2} (useful to see if any computational issues
arose during model fit.)
}
\description{
Return all test statistics values
}
\examples{
fit <- lavaan::sem(txt_mod(1), gen_data_bin(1, n = 500), std.lv = TRUE,
                   estimator = "PML")
all_tests(fit)
}
