% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utilities.R, R/40-test_statistics.R
\name{get_true_values}
\alias{get_true_values}
\alias{get_theoretical_uni_bi_moments}
\title{Returns theoretical true values}
\usage{
get_true_values(model_no)

get_theoretical_uni_bi_moments(model_no, collapse = FALSE)
}
\arguments{
\item{model_no}{(integer) Choose from 1--5. See pkgdown articles for details.}

\item{collapse}{(logical) Should a vector be returned instead of a list
separating the univariate and bivariate quantities?}
}
\value{
A vector of true parameter values used for the simulations.
}
\description{
\itemize{
\item \code{get_true_values()} returns the true values of the freely estimated theta values in this order: \eqn{\lambda} (loadings), \eqn{\rho} (factor correlations), \eqn{\tau} (thresholds).
\item \code{get_theoretical_uni_bi_moments()} returns the univariate (\code{pidot1}) and bivariate (\code{pidot2}) theoretical probabilities of successes.
}
}
\examples{
get_true_values(1)
get_theoretical_uni_bi_moments(1)
}
\seealso{
\code{\link[=gen_data_bin]{gen_data_bin()}}, \code{\link[=gen_data_bin_strat]{gen_data_bin_strat()}},
\code{\link[=gen_data_bin_clust]{gen_data_bin_clust()}}, \code{\link[=gen_data_bin_strcl]{gen_data_bin_strcl()}}, \code{\link[=gen_data_bin_srs]{gen_data_bin_srs()}}
}
