% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.bingof-package.R
\docType{package}
\name{lavaan.bingof-package}
\alias{lavaan.bingof}
\alias{lavaan.bingof-package}
\title{lavaan.bingof: Limited Information Goodness of Fit Tests for Binary Factor Models}
\description{
This is the accompanying R package for the paper "Goodness-of-fit tests for composite likelihood estimation under simple random and complex survey sampling". After fitting a binary factor model in lavaan, the user can call several functions to compute the goodness-of-fit test statistics based off of lower order marginals and compute the corresponding p-values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/haziqj/lavaan.bingof}
  \item \url{http://haziqj.ml/lavaan.bingof/}
  \item Report bugs at \url{https://github.com/haziqj/lavaan.bingof/issues}
}

}
\author{
\strong{Maintainer}: Haziq Jamil \email{haziq.jamil@gmail.com} (\href{https://orcid.org/0000-0003-3298-1010}{ORCID})

}
\keyword{internal}
