% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-test_statistics.R
\name{ligof-test-stats}
\alias{ligof-test-stats}
\alias{Wald_test}
\alias{Wald_test_v2}
\alias{Wald_test_v3}
\alias{Pearson_test_v1}
\alias{Pearson_test_v2}
\alias{RSS_test}
\alias{Multn_test}
\title{Limited information goodness-of-fit test statistics}
\usage{
Wald_test(object, approx_Omega2 = FALSE, svy_design = NULL)

Wald_test_v2(object, approx_Omega2 = FALSE, svy_design = NULL, .order = 3)

Wald_test_v3(object, svy_design = NULL)

Pearson_test_v1(object, approx_Omega2 = FALSE, svy_design = NULL, .order = 2)

Pearson_test_v2(object, approx_Omega2 = FALSE, svy_design = NULL, .order = "3")

RSS_test(object, approx_Omega2 = FALSE, svy_design = NULL, .order = "3")

Multn_test(object, approx_Omega2 = FALSE, svy_design = NULL, .order = "3")
}
\arguments{
\item{object}{A \code{\link[lavaan:lavaan]{lavaan::lavaan()}} fit object.}

\item{approx_Omega2}{(logical) \strong{EXPERIMENTAL} Should an approximate
residual covariance matrix \eqn{\Omega_2} be used? Defaults to \code{FALSE}.}

\item{svy_design}{A \code{\link[survey:svydesign]{survey::svydesign()}} object. Optional.}

\item{.order}{(integer) Either the number of moments to match for the
chi-square test statistic matching procedure (choose from 1--3), or the
Rao-Scott type adjustment order (choose from 1 or 2).}
}
\value{
A data frame containing the test statistics \eqn{X^2}, degrees of
freedom, name of the test, and the \eqn{p}-value.
}
\description{
Limited information goodness-of-fit test statistics
}
\section{Functions}{
\itemize{
\item \code{Wald_test()}: The Wald test statistic.

\item \code{Wald_test_v2()}: The Wald test statistic using a simple diagonal
\eqn{\Omega_2} matrix.

\item \code{Wald_test_v3()}: The Wald test statistic bypassing the
\eqn{\Omega_2} matrix (uses orthogonal complements of \eqn{\Delta_2}).

\item \code{Pearson_test_v1()}: The Pearson test with \eqn{p}-values calculated using a Rao-Scott type adjustment.

\item \code{Pearson_test_v2()}: The Pearson test with \eqn{p}-values calculated using a moment-matching procedure.

\item \code{RSS_test()}: The residual sum of squares (RSS) test. Uses moment-matching for \eqn{p}-value calculations.

\item \code{Multn_test()}: The multinomial test. Uses moment-matching for \eqn{p}-value calculations.

}}
\examples{
fit <- lavaan::sem(txt_mod(1), gen_data_bin(1, n = 500), std.lv = TRUE,
                   estimator = "PML")
Wald_test(fit)
}
\seealso{
\code{\link[=all_tests]{all_tests()}}
}
