% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-sim_functions.R
\name{ligof_sims.methods}
\alias{ligof_sims.methods}
\alias{print.ligof_sims}
\alias{summary.ligof_sims}
\title{Print and summary methods for simulations}
\usage{
\method{print}{ligof_sims}(x, ...)

\method{summary}{ligof_sims}(object, alpha = 0.05, ...)
}
\arguments{
\item{x, object}{The output from \code{\link[=ligof_sims]{ligof_sims()}}.}

\item{...}{Not used.}

\item{alpha}{(numeric) The significance level of the test.}
}
\value{
To be described.
}
\description{
Print and summary methods for simulations
}
