% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-data_gen_complex.R
\name{make_population}
\alias{make_population}
\title{Simulate the school population data set}
\usage{
make_population(
  model_no = 1,
  seed = 123,
  H1 = FALSE,
  return_all = FALSE,
  Sigma2_attr = FALSE
)
}
\arguments{
\item{model_no}{(integer) Choose from 1--5. See pkgdown articles for details.}

\item{seed}{(numeric) The random seed.}

\item{H1}{(logical) Are we generating a data to misspecify the true model?
For power simulations.}

\item{return_all}{(logical) Return the underlying latent variables (\eqn{y^*}
and \eqn{\eta}) as well? well?}

\item{Sigma2_attr}{(logical) Should the population Sigma2 matrix be
computed and stored as an attribute?}
}
\value{
A \code{\link[=tibble]{tibble()}} containing ordinal binary values (0/1) for the items, as
well as the population stratum and clusters (\code{type}, \code{school}, \code{class}).
}
\description{
Simulate the school population data set
}
\examples{
\dontrun{
make_population(1)
}

}
\seealso{
\code{\link[=gen_data_bin_strat]{gen_data_bin_strat()}}, \code{\link[=gen_data_bin_clust]{gen_data_bin_clust()}},
\code{\link[=gen_data_bin_strcl]{gen_data_bin_strcl()}}, \code{\link[=gen_data_bin_srs]{gen_data_bin_srs()}}
}
