% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utilities.R
\name{txt_mod}
\alias{txt_mod}
\title{Function to return the textual model for lavaan fit, depending on which
simulation scenario we are investigating.}
\usage{
txt_mod(model_no = 1L)
}
\arguments{
\item{model_no}{(integer) Choose from 1--5. See pkgdown articles for details.}
}
\value{
Character vector of factor model to input to \code{\link[lavaan:lavaan]{lavaan::lavaan()}}.
}
\description{
Function to return the textual model for lavaan fit, depending on which
simulation scenario we are investigating.
}
\examples{
txt_mod(1)
}
